+++ flowcraft/__init__.py
@@ -1,6 +1,6 @@
 
-__version__ = "1.2.1"
-__build__ = "04072018"
+__version__ = "1.2.2"
+__build__ = "29082018"
 __author__ = "Diogo N. Silva, Tiago F. Jesus, Ines Mendes, Bruno Ribeiro-Goncalves"
 __copyright__ = "Diogo N. Silva"
 __license__ = "GPL3"