+++ flowcraft/generator/components/patlas_mapping.py
@@ -14,13 +14,6 @@ class PatlasMapping(Process):
         self.output_type = "json"
 
         self.params = {
-            "max_k": {
-                "default": 10949,
-                "description": "Sets the k parameter for bowtie2 allowing to "
-                               "make multiple mappings of the same read "
-                               "against several hits on the query sequence or "
-                               "sequences."
-            },
             "trim5": {
                 "default": 0,
                 "description": "Sets trim5 option for bowtie. This will become"
@@ -35,17 +28,17 @@ class PatlasMapping(Process):
                                "covered by reads (in absolute lenght)."
             },
             "refIndex": {
-                "default": "'/ngstools/data/indexes/patlas_bowtie_index'",
+                "default": "'/ngstools/data/indexes/patlas_bowtie2_index'",
                 "description": "Specifies the reference indexes to be provided"
                                " to bowtie2."
             },
             "samtoolsIndex": {
-                "default": "'/ngstools/data/indexes/master_fasta_v1_4_1_11_06_2018.fas.fai'",
+                "default": "'/ngstools/data/indexes/master_fasta_plasmid_db.fas.fai'",
                 "description": "Specifies the reference indexes to be provided"
                                " to samtools."
             },
             "lengthJson": {
-                "default": "'/ngstools/data/length_v1_4_1_11_06_2018.json'",
+                "default": "'/ngstools/data/length_plasmid_db.json'",
                 "description": "A dictionary of all the lengths of reference "
                                "sequences."
             }
@@ -54,19 +47,14 @@ class PatlasMapping(Process):
         self.directives = {
             "mappingBowtie": {
                 "container": "flowcraft/mapping-patlas",
-                "version": "1.4.1",
+                "version": "1.5.2-2",
                 "cpus": 1,
-                "memory": "{ 4.GB * task.attempt }"
-            },
-            "samtoolsView": {
-                "container": "flowcraft/mapping-patlas",
-                "version": "1.4.1",
-                "cpus": 1,
-                "memory": "{ 4.GB * task.attempt }"
+                "memory": "{ 4.GB * task.attempt }",
+                "scratch": "true"
             },
             "jsonDumpingMapping": {
                 "container": "flowcraft/mapping-patlas",
-                "version": "1.4.1",
+                "version": "1.5.2-2",
                 "cpus": 1,
                 "memory": "'4GB'"
             }
@@ -74,7 +62,6 @@ class PatlasMapping(Process):
 
         self.status_channels = [
             "mappingBowtie",
-            "samtoolsView",
             "jsonDumpingMapping"
         ]