+++ flowcraft/generator/recipe.py
@@ -569,4 +569,5 @@ available_recipes = {
     "plasmids_assembly": "integrity_coverage fastqc_trimmomatic (spades pilon"
                          " mash_dist)",
     "plasmids_mash": "integrity_coverage fastqc_trimmomatic mash_screen",
+    "den-im": "integrity_coverage fastqc_trimmomatic filter_poly remove_host bowtie retrieve_mapped viral_assembly assembly_mapping pilon split_assembly (dengue_typing | mafft raxml)",
 }